requirements.txt
ctransformers
sentence-transformers
pinecone-client
openai
langchain
langchain-community
flask
pypdf
python-dotenv
-e .